much better setup

+++ packages/core/src/infrastructure/database/shared/migrate.ts
@@ -1,9 +1,20 @@
 import { promises as fs } from 'fs';
-import { FileMigrationProvider, Migrator } from 'kysely';
+import { FileMigrationProvider, Kysely, Migrator } from 'kysely';
+import { DB } from 'kysely-codegen/dist/db';
 import path from 'path';
 import { fileURLToPath } from 'url';
 
 import { createDatabaseConnection } from './create-database-connection';
+
+type MigrateOptions = {
+  db?: Kysely<DB>;
+  down?: boolean;
+};
+
+const defaultOptions: MigrateOptions = {
+  db: undefined,
+  down: false,
+};
 
 /**
  * Migrates the database to the latest version by executing all migrations.
@@ -14,11 +25,16 @@ import { createDatabaseConnection } from './create-database-connection';
  * we want to ensure that the database is migrated to the latest version before
  * we seed it.
  */
-export async function migrate(down: boolean = false) {
+export async function migrate(options: MigrateOptions = defaultOptions) {
+  options = {
+    ...defaultOptions,
+    ...options,
+  };
+
+  const db = options.db || createDatabaseConnection();
+
   const __filename = fileURLToPath(import.meta.url);
   const __dirname = path.dirname(__filename);
-
-  const db = createDatabaseConnection();
 
   const migrator = new Migrator({
     db,
@@ -31,7 +47,7 @@ export async function migrate(down: boolean = false) {
     migrationLockTableName: 'kysely_migrations_lock',
   });
 
-  const { error, results } = !!down
+  const { error, results } = !!options.down
     ? await migrator.migrateDown()
     : await migrator.migrateToLatest();
 
@@ -56,5 +72,9 @@ export async function migrate(down: boolean = false) {
     process.exit(1);
   }
 
-  await db.destroy();
+  // If a database instance was passed in, we'll yield the responsibility of
+  // destroying it to the caller. Otherwise, we'll destroy it here.
+  if (!options.db) {
+    await db.destroy();
+  }
 }
